Fix float precision loss in PriceService discount calculation
discountPercent / 100f forced the base price through float, which
loses integer precision above 2^24 and could return a discounted
price off by whole units for large base prices. Switched the
calculation to double.

+++ src/main/java/com/example/PriceService.java
@@ -28,7 +28,7 @@ public class PriceService {
                     "discountPercent must be 0-100, got " + discountPercent);
         }
         int base = client.fetchPrice(sku);
-        int discount = Math.round(base * (discountPercent / 100f));
+        int discount = (int) Math.round(base * (discountPercent / 100.0));
         return Math.max(0, base - discount);
     }
